Implementa CRUD de Usuario e Perfil
Implementação inicial do CRUD de usuario e perfil.

+++ api/usuario-service/src/main/java/br/com/softplan/repository/PerfilRepository.java
@@ -3,5 +3,5 @@ package br.com.softplan.repository;
 import br.com.softplan.entidades.Perfil;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface PerfilRepository extends PagingAndSortingRepository<Perfil, Integer> {
+public interface PerfilRepository extends PagingAndSortingRepository<Perfil, Long> {
 }
