+++ api/usuario-service/src/main/java/br/com/softplan/repository/UsuarioRepository.java
@@ -2,6 +2,8 @@ package br.com.softplan.repository;
 
 import br.com.softplan.entidades.Usuario;
 import org.springframework.data.repository.PagingAndSortingRepository;
+import org.springframework.data.repository.query.Param;
 
-public interface UsuarioRepository extends PagingAndSortingRepository<Usuario, Integer> {
+public interface UsuarioRepository extends PagingAndSortingRepository<Usuario, Long> {
+
 }
